quiet some warnings which appear on vs2015.

+++ Source/Core/Core/PowerPC/Jit64/Jit_Integer.cpp
@@ -1525,7 +1525,7 @@ void Jit64::rlwimix(UGeckoInstruction inst)
 			bool isRightShift = mask == (1U << inst.SH) - 1;
 			if (gpr.R(a).IsImm())
 			{
-				u32 maskA = gpr.R(a).offset & ~mask;
+				u32 maskA = (u32)gpr.R(a).offset & ~mask;
 				gpr.BindToRegister(a, false, true);
 				MOV(32, gpr.R(a), gpr.R(s));
 				if (isLeftShift)
